AthenaColumnGetterFactory move to src/main/java/org/embulk/input/athena/getter/
