+++ src/main/java/org/embulk/input/athena/AthenaInputPlugin.java
@@ -30,6 +30,7 @@ import org.embulk.config.ConfigException;
 import org.embulk.config.ConfigSource;
 import org.embulk.config.TaskReport;
 import org.embulk.config.TaskSource;
+import org.embulk.input.athena.getter.AthenaColumnGetterFactory;
 import org.embulk.input.jdbc.getter.ColumnGetterFactory;
 import org.embulk.input.jdbc.JdbcColumn;
 import org.embulk.input.jdbc.JdbcColumnOption;
@@ -326,29 +327,10 @@ public class AthenaInputPlugin implements InputPlugin
 
         return DriverManager.getConnection(task.getAthenaUrl(), properties);
     }
-
-
+    
     private ColumnGetterFactory newColumnGetterFactory(PageBuilder pageBuilder, ZoneId dateTimeZone)
     {
         return new AthenaColumnGetterFactory(pageBuilder, dateTimeZone);
-    }
-
-    private static class AthenaColumnGetterFactory extends ColumnGetterFactory {
-        public AthenaColumnGetterFactory(PageBuilder to, ZoneId defaultTimeZone) {
-            super(to, defaultTimeZone);
-        }
-
-        @Override
-        protected String sqlTypeToValueType(JdbcColumn column, int sqlType) {
-            try {
-                return super.sqlTypeToValueType(column, sqlType);
-            } catch (UnsupportedOperationException e) {
-                throw new UnsupportedOperationException(
-                        String.format(ENGLISH,
-                                "Unsupported type %s (sqlType=%d) of '%s' column. Please add '%s: {value_type: string}' to 'column_options: {...}' option to convert the values to strings.",
-                                column.getTypeName(), column.getSqlType(), column.getName(), column.getName()));
-            }
-        }
     }
 
     //
